MSP-2288: Run slurm jobs in cgroups + fix release versioning
X-Arc-Original-Commit: a03e532241c5bd5afe2452b908a34971d0111d1f

+++ images/docker/populate_jail/populate_jail_entrypoint.sh
@@ -5,6 +5,11 @@ while ! mountpoint -q /mnt/jail; do
     sleep 10
 done
 
+if [ "$OVERWRITE" != "1" ] && [ -d /mnt/jail/dev ]; then
+    echo "Jail is already populated and content overwriting is turned off, exiting"
+    exit 0
+fi
+
 echo "Delete everything from jail directory"
 rm -rf /mnt/jail/*
